Move files into security folder
