+++ src/main/java/com/unblock/server/security/JWTLoginFilter.java
@@ -1,6 +1,8 @@
-package com.unblock.server;
+package com.unblock.server.security;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.unblock.server.security.AccountCredentials;
+import com.unblock.server.security.TokenAuthenticationService;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
@@ -13,7 +15,6 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.Collections;
 
 public class JWTLoginFilter extends AbstractAuthenticationProcessingFilter {
